backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
djongo==1.3.6
pymongo==4.6.1
python-telegram-bot==21.9
instagrapi==2.1.2
python-dotenv==1.0.1
requests==2.32.3
Pillow==11.1.0